mid-project-347359563/services/booking-ticket-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-dotenv
pika
httpx
aio-pika
requests
cryptography